Update Analyzer.py
Fixed issue where Analyzer required a Detoxigramer object for initialization. This was incorrect since Detoxigramers change with each bot interaction due to varying chat contexts.

The solution is to pass a ManagementDetoxigramers() object, which manages all users. Functions now require a user_id, generated by hashing the phone number or wa-id (for WhatsApp) with SHAKE-256.

This is achieved when setting an id, and will always be the same id.